fix: lower pulse worker status complexity
